Learning companion/requirements.txt
streamlit
requests
sentence-transformers
faiss-cpu
python-dotenv
pillow
pytesseract
pdfplumber 